add service alert call to api interface, refactor bustime response to include service alert
