+++ app/src/main/java/com/taitsmith/busboy/api/ApiInterface.kt
@@ -56,6 +56,13 @@ interface ApiInterface {
         @Query("token") token: String = BuildConfig.ac_transit_key
      ): List<Bus>
 
+     //get alerts for lines served by selected stop
+     @GET("actrealtime/servicebulletin")
+     suspend fun getServiceAlertsForStop(
+         @Query("rt") route: String,
+         @Query("token") token: String = BuildConfig.ac_transit_key
+     ): List<ServiceAlert>
+
     //talk to google and get walking directions from our location to the selected stop
     @GET("maps/api/directions/json")
      suspend fun getNavigationToStop(
