+++ app/src/main/java/com/taitsmith/busboy/api/BustimeResponse.kt
@@ -1,4 +1,41 @@
 package com.taitsmith.busboy.api
 
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+import com.taitsmith.busboy.data.Prediction
+import com.taitsmith.busboy.data.ServiceAlert
+
+/**
+ * bustime-response is a general class returned from ac transit prediction api calls as well as
+ * ac transit service alert api calls.
+ */
 class BustimeResponse {
+    //body returned for stop prediction calls
+    @SerializedName("prd")
+    @Expose
+    val prd: List<Prediction>? = null
+
+    //body returned for service alert calls
+    @SerializedName("sb")
+    @Expose
+    val sb: List<ServiceAlert>? = null
+
+    //almost always empty
+    @SerializedName("error")
+    @Expose
+    val error: List<BusError>? = null
+
+    class BusError {
+        @SerializedName("rtpidatafeed")
+        @Expose
+        var rtpidatafeed: String? = null
+
+        @SerializedName("stpid")
+        @Expose
+        var stpid: String? = null
+
+        @SerializedName("msg")
+        @Expose
+        var msg: String? = null
+    }
 }
